Added the last piece, the feed table view controller. Updates from the parse database. Textfields load even if images have not downloaded yet.
This is the end of the Instagram Clone exercise.

Things to note (apart from the 10MB limit I encountered) is that this app would not work with a huge load of users (think 1000s or more) since it downloads all the users and data at once. Additional data can easily be added, such as date of post entry, location and other things that might be of interest to the post.

What I learned from this exercise... copy notes :P Well, not exactly true. I had to problem solve when things went south.